requirements.txt
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.0.12
Django==4.0.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
pycodestyle==2.8.0
PyJWT==2.4.0
pytz==2022.1
requests==2.28.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.9
